# server/api-docs.yaml
openapi: 3.0.3
info:
  title: Weather Sunset Predictor API - 天气晚霞预测器后端 API
  description: |
    天气晚霞预测器（Weather Sunset Predictor）后端 API 文档。

    本系统基于 Windy Point Forecast API 提供实时天气数据代理，并实现了多因子火烧云（晚霞/朝霞）
    预测算法，包括基础单点预测、增强版物理模型预测、周边8方向聚合预测、以及基于 GFS 气象数据
    的火烧云地图覆盖层生成。

    ## 架构说明
    - 后端作为代理层，保护 Windy API 密钥不暴露给前端
    - 预测算法在后端运行，支持缓存和多客户端复用
    - 火烧云覆盖层通过 Python 脚本处理 NOAA GFS GRIB2 数据生成

    ## 缓存策略
    | 数据类型 | TTL |
    |---------|-----|
    | 预测数据 | 30 分钟 |
    | 周边预测 | 1 小时 |
    | 天气数据 | 15 分钟 |
    | 覆盖层数据 | 30 分钟 |

    ## 认证方式
    API 密钥通过服务器端环境变量（`WINDY_API_KEY`、`WINDY_MAP_API_KEY`）配置，
    前端客户端无需提供 API 密钥，直接调用后端代理即可。

    最后更新：2026-02-11（补充天气代理、系统配置与火烧云 API 端点示例）
  version: 1.1.0
  contact:
    name: Weather Sunset Predictor
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: 本地开发服务器（默认端口 3000）
...
paths:
...
  /api/prediction/enhanced:
    post:
      tags:
        - 预测API
      summary: 增强版单点火烧云预测
      description: |
        增强版火烧云预测算法（Phase 3），基于物理模型的四模块评分系统：

        1. **时间判定逻辑**（触发机制）：根据太阳高度角判断是否在火烧云观测窗口内
        2. **画布逻辑**（本地云况评分）：评估高/中/低云的加权有效云量，低云惩罚
        3. **光路逻辑**（光路通透评分）：评估 150km 和 300km 远处的云量，判断光路是否通畅
        4. **渲染逻辑**（画质修正系数）：能见度、湿度、AQI、雨后初晴等修正因子

        综合得分 = (画布 x 0.4 + 光路 x 0.6) x 渲染修正系数

        状态判定结果包括：`no_fire_cloud`、`light_glow`、`good_glow`、`very_likely`、`legendary_eruption`
      operationId: calculateEnhancedPrediction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnhancedPredictionRequest'
            example:
              weatherData:
                lowClouds: 15
                midClouds: 50
                highClouds: 35
                visibility: 20
                humidity: 45
                aqi: 40
              date: "2026-06-21T18:00:00Z"
              lat: 39.9042
              lon: 116.4074
              type: sunset
              options:
                remoteCloudData:
                  near:
                    totalCloud: 30
                  far:
                    totalCloud: 20
                rainedRecently: false
      responses:
        '200':
          description: 成功返回增强预测结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnhancedPredictionResponse'
              example:
                success: true
                data:
                  date: "2026-06-21T18:00:00.000Z"
                  type: sunset
                  score: 72.5
                  quality: good
                  status: good_glow
                  icon: city_sunset
                  description: conditions_good
                  advice: can_watch
                  timeAnalysis:
                    inWindow: true
                    elevation: -0.83
                    optimalMoment: mid_cloud_eruption
                    windowDescription: "solar_elevation:-0.8"
                    isInDetectionWindow: true
                  breakdown:
                    baseScore: 72.5
                    canvasScore: 65.3
                    lightPathScore: 77.3
                    renderingFactor: 1.0
                  canvasAnalysis:
                    score: 65.3
                    effectiveCloudCover: 38.5
                    cloudRangeScore: 76.4
                    cloudLevel: perfect
                    lowCloudPenalty: 0.85
                    penaltyReason: no_low_cloud_obstruction
                    penaltyValue: 15
                    breakdown:
                      lowClouds: 15.0
                      midClouds: 50.0
                      highClouds: 35.0
                  lightPathAnalysis:
                    score: 77.3
                    nearPointScore: 71.4
                    farPointScore: 85.7
                    breakdown:
                      nearPointCloudCover: 30
                      farPointCloudCover: 20
                    hasRemoteData: true
                  renderingAnalysis:
                    factor: 1.0
                    visibilityFactor: 1.1
                    humidityFactor: 1.0
                    rainBonus: 1.0
                    breakdown:
                      visibility: excellent
                      humidity: moderate
                      aqi: excellent
                      colorTendency: golden_orange
                      specialMode: null
        '400':
          description: 请求参数验证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: 增强预测计算出错
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
              example:
                error: PREDICTION_ERROR
                message: "计算失败"
...
  /api/agent/openapi.json:
    get:
      tags:
        - Agent API
      summary: 获取 Agent OpenAPI 规格
      description: 获取 Agent 相关接口文档（公开访问，受速率限制）。
      responses:
        '200':
          description: 成功返回 OpenAPI JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenApiSpec'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/agent/forecast:
    get:
      tags:
        - Agent API
      summary: 火烧云预测
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: lat
          schema:
            type: number
          required: false
          description: 纬度（-90 到 90）
        - in: query
          name: lon
          schema:
            type: number
          required: false
          description: 经度（-180 到 180）
        - in: query
          name: location
          schema:
            type: string
          required: false
          description: 地点名称，lat/lon 与其二选一
        - in: query
          name: type
          schema:
            type: string
            enum: [sunrise, sunset]
          required: false
          description: 预测时段
        - in: query
          name: date
          schema:
            type: string
          required: false
          description: today / tomorrow / YYYY-MM-DD
        - in: query
          name: detail
          schema:
            type: string
            enum: [simple, full]
          required: false
          description: 响应详细度
      responses:
        '200':
          description: 预测结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentForecastResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/agent/explain:
    get:
      tags:
        - Agent API
      summary: 预测解释
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: lat
          required: false
          schema:
            type: number
          description: 纬度
        - in: query
          name: lon
          required: false
          schema:
            type: number
          description: 经度
        - in: query
          name: location
          required: false
          schema:
            type: string
          description: 地点名称，lat/lon 与其二选一
        - in: query
          name: type
          required: false
          schema:
            type: string
            enum: [sunrise, sunset]
          description: 预测时段
        - in: query
          name: date
          required: false
          schema:
            type: string
          description: today / tomorrow / YYYY-MM-DD
      responses:
        '200':
          description: 解释结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentExplainResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/agent/geocode:
    get:
      tags:
        - Agent API
      summary: 地理编码
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
          description: 查询关键词
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
          description: 返回条数
      responses:
        '200':
          description: 地理编码结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentGeocodeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalError'


# ==================== 数据模型定义 ====================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
  schemas:
    AgentForecastResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            location:
              type: object
            type:
              type: string
            date:
              type: string
            score:
              type: number
            quality:
              type: string
            bestViewingWindow:
              type: object
      required:
        - success
        - data
    AgentExplainResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            location:
              type: object
            factorRelations:
              type: array
              items:
                type: object
            constraints:
              type: array
              items:
                type: object
            scoreComposition:
              type: object
            explanation:
              type: object
      required:
        - success
        - data
    AgentGeocodeResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            query:
              type: string
            top:
              type: object
            results:
              type: array
              items:
                type: object
      required:
        - success
        - data
    OpenApiSpec:
      type: object
      description: Agent OpenAPI 文档 JSON

    # ---------- 通用响应 ----------
...
    ValidationErrorResponse:
      type: object
      description: 预测 API 验证错误响应格式
      properties:
        error:
          type: string
          description: 错误代码
        message:
          type: string
          description: 错误描述信息
      required:
        - error
        - message

    # ---------- 天气数据 ----------
...
    EnhancedWeatherDataInput:
      type: object
      description: 增强预测的天气数据输入
      properties:
        lowClouds:
          type: number
          description: 低云覆盖率（百分比，0-100）
        midClouds:
          type: number
          description: 中云覆盖率（百分比，0-100）
        highClouds:
          type: number
          description: 高云覆盖率（百分比，0-100）
        visibility:
          type: number
          description: 能见度（km）
        humidity:
          type: number
          description: 相对湿度（百分比，0-100）
        aqi:
          type: number
          description: 空气质量指数（AQI，默认 50）
    EnhancedPredictionRequest:
      type: object
      description: 增强版单点火烧云预测请求
      properties:
        weatherData:
          $ref: '#/components/schemas/EnhancedWeatherDataInput'
        date:
          type: string
          format: date-time
          description: 预测日期（ISO 8601 格式）
        lat:
          type: number
          format: double
          minimum: -90
          maximum: 90
          description: 纬度
        lon:
          type: number
          format: double
          minimum: -180
          maximum: 180
          description: 经度
        type:
          type: string
          enum:
            - sunrise
            - sunset
          description: 预测类型
        options:
          $ref: '#/components/schemas/EnhancedPredictionOptions'
      required:
        - weatherData
        - date
        - lat
        - lon
        - type
...
    EnhancedPredictionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/EnhancedPredictionResult'
      required:
        - success
        - data

    # ---------- 批量预测 ----------
...